Rename EventHelper to EventFactory. Move print method to EventExtentions.
