+++ tools/Custom/EventFactory.cs
@@ -10,7 +10,7 @@ using Microsoft.Graph.PowerShell.Runtime;
 
 namespace Microsoft.Graph.PowerShell
 {
-    public static class EventHelper
+    public static class EventFactory
     {
         /// <summary>
         /// Create a tracing event containing a string message
@@ -53,30 +53,6 @@ namespace Microsoft.Graph.PowerShell
                 Message = message
             };
         }
-
-        /// <summary>
-        /// Print event details to the provided stream
-        /// </summary>
-        /// <param name="getEventData">The event data to print</param>
-        /// <param name="signal">The delegate for signaling events to the runtime</param>
-        /// <param name="token">The cancellation token for the request</param>
-        /// <param name="streamName">The name of the stream to print data to</param>
-        /// <param name="eventName">The name of the event to be printed</param>
-        public static async void Print(this Func<EventArgs> getEventData, Func<string, CancellationToken, Func<EventArgs>, Task> signal, CancellationToken token, string streamName, string eventName)
-        {
-            var eventDisplayName = SplitPascalCase(eventName).ToUpperInvariant();
-            var data = EventDataConverter.ConvertFrom(getEventData()); // also, we manually use our TypeConverter to return an appropriate type
-            if (data.Id != "Verbose" && data.Id != "Warning" && data.Id != "Debug" && data.Id != "Information" && data.Id != "Error")
-            {
-                await signal(streamName, token, () => EventHelper.CreateLogEvent($"{eventDisplayName} The contents are '{data?.Id}' and '{data?.Message}'"));
-                if (data != null)
-                {
-                    await signal(streamName, token, () => EventHelper.CreateLogEvent($"{eventDisplayName} Parameter: '{data.Parameter}'\n{eventDisplayName} RequestMessage '{data.RequestMessage}'\n{eventDisplayName} Response: '{data.ResponseMessage}'\n{eventDisplayName} Value: '{data.Value}'"));
-                    await signal(streamName, token, () => EventHelper.CreateLogEvent($"{eventDisplayName} ExtendedData Type: '{data.ExtendedData?.GetType()}'\n{eventDisplayName} ExtendedData '{data.ExtendedData}'"));
-                }
-            }
-        }
-
         static string SplitPascalCase(string word)
         {
             var regex = new Regex("([a-z]+)([A-Z])");
